Lagt til removal of session
